Add incomplete python grammar and meta-grammar parser
The meta-grammar parser in `TyPy.Compiler\peg.py` parses the peg format for the python grammar. The meta-grammar parser is a poor front end and cannot perform complex diagnostics. This uses the current python grammar and a lookup table to generate the relevant C# code that represents our parser in IParsable format. It is important to note that the python grammar has elements of left-recursion which will need to be manually removed. The next step will be creating the relevant algorithms for detecting left recursion and factorising left recursion out in basic cases. We are aware that this will mean that there is a mismatch between the internal trees produced, but this should not be an issue.

A placeholder `Cut` `IParsable` is put in place until we can sufficiently support this in the grammar.

Other helper classes have been created in order to more easily match the python grammar.